making modularized code splitting code into multiple py files for easy debugging and maintenance
