+++ Project2/gui.py
@@ -1,0 +1,1 @@
+# Placeholder for GUI-related code
